tests/integration: Move test "ports" to corresponding directory
