+++ tests/integration/test_podman_compose_tests.py
@@ -16,34 +16,6 @@ from tests.integration.test_utils import RunSubprocessMixin
 
 
 class TestPodmanCompose(unittest.TestCase, RunSubprocessMixin):
-    def test_up_with_ports(self):
-        up_cmd = [
-            "coverage",
-            "run",
-            podman_compose_path(),
-            "-f",
-            os.path.join(test_path(), "ports", "docker-compose.yml"),
-            "up",
-            "-d",
-            "--force-recreate",
-        ]
-
-        down_cmd = [
-            "coverage",
-            "run",
-            podman_compose_path(),
-            "-f",
-            os.path.join(test_path(), "ports", "docker-compose.yml"),
-            "down",
-            "--volumes",
-        ]
-
-        try:
-            self.run_subprocess_assert_returncode(up_cmd)
-
-        finally:
-            self.run_subprocess_assert_returncode(down_cmd)
-
     def test_down_with_vols(self):
         up_cmd = [
             "coverage",
@@ -83,42 +55,3 @@ class TestPodmanCompose(unittest.TestCase, RunSubprocessMixin):
             self.run_subprocess(["podman", "volume", "rm", "actual-name-of-volume"])
             self.assertEqual(return_code, 0)
 
-    def test_down_with_orphans(self):
-        container_id, _ = self.run_subprocess_assert_returncode([
-            "podman",
-            "run",
-            "--rm",
-            "-d",
-            "nopush/podman-compose-test",
-            "dumb-init",
-            "/bin/busybox",
-            "httpd",
-            "-f",
-            "-h",
-            "/etc/",
-            "-p",
-            "8000",
-        ])
-
-        down_cmd = [
-            "coverage",
-            "run",
-            podman_compose_path(),
-            "-f",
-            os.path.join(test_path(), "ports", "docker-compose.yml"),
-            "down",
-            "--volumes",
-            "--remove-orphans",
-        ]
-
-        self.run_subprocess_assert_returncode(down_cmd)
-
-        self.run_subprocess_assert_returncode(
-            [
-                "podman",
-                "container",
-                "exists",
-                container_id.decode("utf-8"),
-            ],
-            1,
-        )
